+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## Next
 
+## Added
+- Tint color ([#15](https://github.com/AckeeCZ/ACKImagePicker/pull/15)) by @LukasHromadnik
+
 ## 0.3.0
 
 ## Added